fixup! fixup! refactor: mock response with pytest_httpx Refs: 208
