+++ tests/extensions/test_langchain.py
@@ -9,8 +9,8 @@ from genai import Credentials
 from genai.schemas import GenerateParams
 from genai.schemas.responses import GenerateResponse, GenerateStreamResponse
 from genai.services import ServiceInterface
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 
 @pytest.mark.extension
