+++ tests/test_generate_service.py
@@ -6,8 +6,8 @@ from pytest_httpx import HTTPXMock
 
 from genai.schemas import GenerateParams
 from genai.services import ServiceInterface
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 
 @pytest.mark.unit
