+++ tests/test_metadata.py
@@ -4,8 +4,8 @@ from pytest_httpx import HTTPXMock
 from genai import Credentials, Metadata
 from genai.schemas.responses import HistoryResponse, TermsOfUse
 from genai.services import ServiceInterface
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 
 @pytest.mark.unit
