+++ tests/test_model_async.py
@@ -11,8 +11,8 @@ from genai.exceptions import GenAiException
 from genai.schemas import GenerateParams, TokenParams
 from genai.schemas.responses import GenerateResponse, GenerateResult, TokenizeResponse
 from genai.services import ServiceInterface
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 
 @pytest.mark.unit
