+++ tests/test_service_interface_async.py
@@ -6,8 +6,8 @@ from pytest_httpx import HTTPXMock
 from genai.exceptions import GenAiException
 from genai.services import ServiceInterface
 from genai.services.connection_manager import ConnectionManager
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 # API Reference : https://workbench.res.ibm.com/docs/api-reference#generate
