+++ tests/test_tunes.py
@@ -19,8 +19,8 @@ from genai.schemas.tunes_params import (
     TunesListParams,
 )
 from genai.services.tune_manager import TuneManager
-from genai.utils.request_utils import match_endpoint
 from tests.assets.response_helper import SimpleResponse
+from tests.utils import match_endpoint
 
 
 @pytest.mark.unit
