+++ tests/test_utils.py
@@ -1,7 +1,10 @@
+from re import Match
+
 import pytest
 from pydantic import BaseModel, ValidationError
 
 from genai.utils.general import to_model_instance
+from tests.utils import match_endpoint
 
 
 @pytest.mark.unit
@@ -20,3 +23,12 @@ class TestUtils:
             to_model_instance([], Company)
         with pytest.raises(ValidationError):
             to_model_instance({}, Company)
+
+    def test_match_endpoint(self):
+        match_ep = match_endpoint("v1/generate")
+        url = "http://service_url/v1/tokenize"
+        assert match_ep.match(url) is None
+
+        match_ep = match_endpoint("v1/generate", "path", query_params={"parameter": 20, "test": "test"})
+        url = "http://service_url/v1/generate/path?parameter=20&test=test"
+        assert isinstance(match_ep.match(url), Match) is True
